Change to MIT license
Unsure whether GPL3 would stop someone from re-using this code and/or contributing to it. Therefore switching to MIT. 
The main goal is that the code can help ABAP developers handle SAPScript texts. MIT will allow everyone to do that, even if they use it in a product they sell.